+++ server/tests/api/test_states.py
@@ -3,14 +3,9 @@
 
 
 import asyncio
-import datetime
-import json
-import time
 import uuid
 
-import pendulum
 import pytest
-from asynctest import CoroutineMock
 from box import Box
 
 import prefect