+++ src/prefect/client/secrets.py
@@ -56,6 +56,7 @@ from typing import Any, Optional
 
 import prefect
 from prefect.client.client import Client
+from prefect.utilities.exceptions import ClientError
 
 
 class Secret:
@@ -113,7 +114,8 @@ class Secret:
         Raises:
             - ValueError: if `.get()` is called within a Flow building context, or if `use_local_secrets=True`
                 and your Secret doesn't exist
-            - ClientError: if `use_local_secrets=False` and the Client fails to retrieve your secret
+            - KeyError: if `use_local_secrets=False` and the Client fails to retrieve your secret
+            - ClientError: if the client experiences an unexpected error communicating with the backend
         """
         if isinstance(prefect.context.get("flow"), prefect.core.flow.Flow):
             raise ValueError(
@@ -125,14 +127,22 @@ class Secret:
             value = secrets[self.name]
         except KeyError:
             if prefect.context.config.cloud.use_local_secrets is False:
-                result = self.client.graphql(
-                    """
-                    query($name: String!) {
-                        secret_value(name: $name)
-                    }
-                    """,
-                    variables=dict(name=self.name),
-                )
+                try:
+                    result = self.client.graphql(
+                        """
+                        query($name: String!) {
+                            secret_value(name: $name)
+                        }
+                        """,
+                        variables=dict(name=self.name),
+                    )
+                except ClientError as exc:
+                    if "No value found for the requested key" in str(exc):
+                        raise KeyError(
+                            f"The secret {self.name} was not found.  Please ensure that it was set correctly in your tenant: https://docs.prefect.io/orchestration/concepts/secrets.html"
+                        )
+                    else:
+                        raise exc
                 # the result object is a Box, so we recursively restore builtin
                 # dict/list classes
                 result_dict = result.to_dict()