+++ tests/engine/executors/test_executors.py
@@ -1,13 +1,10 @@
-import datetime
 import logging
 import random
 import sys
-import tempfile
 import time
 from unittest.mock import MagicMock
 
 import cloudpickle
-import dask
 import distributed
 import pytest